Rename operator API header file to operator_api.h
