+++ examples/c-operator/README.md
@@ -3,7 +3,7 @@
 Build with these steps:
 
 ```bash
-cp ../../../apis/c/operator/api.h .
+cp ../../apis/c/operator/operator_api.h .
 clang -c operator.c
 clang -shared -v operator.o -o operator.so
 ```
