+++ examples/c-operator/operator.c
@@ -1,4 +1,4 @@
-#include "api.h"
+#include "operator_api.h"
 #include <string.h>
 #include <stdio.h>
 #include <stdlib.h>
